+++ src/app/shared/config/firebase.config.ts
@@ -1,0 +1,8 @@
+export const firebaseConfig = {
+  apiKey: 'AIzaSyBkkjVmAcq4CZRkr-Oc0lcCGKfSqSZRjdQ',
+  authDomain: 'venture-map-web.firebaseapp.com',
+  projectId: 'venture-map-web',
+  storageBucket: 'venture-map-web.firebasestorage.app',
+  messagingSenderId: '457351831227',
+  appId: '1:457351831227:web:c2bb8010bf390f49cf939e',
+};